+++ Tests/Fixtures/Model/ServiceStub.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the "schema" extension for TYPO3 CMS.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ */
+
+namespace Brotkrueml\Schema\Tests\Fixtures\Model;
+
+class ServiceStub extends GenericStub
+{
+    public function __construct()
+    {
+        parent::__construct(
+            null,
+            [
+                'name' => null,
+                'serviceType' => null,
+            ],
+            'ServiceStub'
+        );
+    }
+}